Move stackbit page to vital-test
