+++ packages/vital-templates/src/components/Page/index.ts
@@ -15,5 +15,5 @@
 export { default as vitalPage } from './tokens';
 export { default as vitalPageBase } from './tokens/vitalPage';
 
-export type { PageProps } from './asBodilessPage';
-export { asBodilessPage } from './asBodilessPage';
+export type { PageProps } from '../../util/asBodilessPage';
+export { asBodilessPage } from '../../util/asBodilessPage';
