+++ packages/vital-test/src/shadow/@bodiless/vital-templates/Page.ts
@@ -1,0 +1,14 @@
+import vitalPageBase from '--vital--/lib/shadow/@bodiless/vital-templates/Page';
+import { asFluidToken } from '@bodiless/vital-elements';
+import { asBodilessPage } from '../../../helpers/asBodilessPage';
+
+const Default = asFluidToken({
+  ...vitalPageBase.Default,
+  Schema: {
+    _: asBodilessPage,
+  },
+});
+
+export default {
+  Default,
+};
